Migrate Gcp-related secret unit tests.
